Extract common doc uploading into DocumentUploadQuestionController
This new controller introduces a new subclass of QuestionsController
which contains the general logic for a document upload page. It requires
subclasses to implement the `document_type` method, and everything else
should just work :sparkles:.
