+++ app/controllers/questions/w2s_controller.rb
@@ -1,24 +1,10 @@
 module Questions
-  class W2sController < QuestionsController
-    layout "document_upload"
+  class W2sController < DocumentUploadQuestionController
 
-    def section_title
-      "Documents"
-    end
+    private
 
-    def edit
-      @documents = current_intake.documents.of_type("W-2")
-      super
-    end
-
-    def update
-      @form = form_class.new(current_intake, form_params)
-      if @form.valid?
-        @form.save
-        update_session
-      end
-
-      redirect_to action: :edit
+    def document_type
+      "W-2"
     end
   end
 end
