+++ lib/generators/templates/document_upload_form.template
@@ -1,11 +1,13 @@
+# frozen_string_literal: true
+
 class <%= @form_name %> < QuestionsForm
   set_attributes_for :intake, :document
 
   def save
     document_file_upload = attributes_for(:intake)[:document]
-    if document_file_upload.present?
-      document = @intake.documents.create(document_type: <%= options.document_type.inspect %>)
-      document.upload.attach(document_file_upload)
-    end
+    return unless document_file_upload.present?
+
+    document = @intake.documents.create(document_type: <%= options.document_type.inspect %>)
+    document.upload.attach(document_file_upload)
   end
 end
